backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0
pydantic[email]>=2.0,<3.0
pydantic-settings==2.7.1
sqlalchemy>=2.0,<3.0
asyncpg==0.30.0
alembic==1.14.1
redis[hiredis]==5.2.1
upstash-redis==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
boto3==1.36.7
structlog==24.4.0
sentry-sdk[fastapi]==2.19.2
opentelemetry-sdk==1.29.0
opentelemetry-api==1.29.0
prometheus-client==0.21.1
langchain==0.3.15
langgraph==0.2.62
anthropic==0.42.0
openai==1.59.7
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.2
faker==33.3.1
cryptography==44.0.0
